Initial commit
This change introduces a Prometheus exporter to publish
ECS metrics in the Prometheus exposition format.

The ecs_exporter reads the ECS metadata server to read
ECS task and container stats. This is why
the exporter needs to run as a sidecar container
in an ECS task, or it should be included as a binary in
a container deployed to be able to fetch from the
metadata server.

The exporter publishes metrics from the container runtime
related to CPU usage, networking and more.
Users still need to publish their custom application
metrics from their own containers if they have any.

Currently, these metrics are available from CloudWatch and
can be ingested into Prometheus by using the cloudwatch_exporter,
but there are various users who want to be able to directly
export them to Prometheus to avoid delay in data collection.

In the future, the exporter will report more CPU, memory,
and I/O metrics.

Fixes prometheus-community/community#36.

Signed-off-by: JBD <jbd@amazon.com>