Network metrics: drop `container_name` label (#95)
I made a mistake in #81, incorrectly determining that network stats were
scoped to containers as well as network interfaces. They are not, in
spite of being laid out that way in the API response. And my tests
consisted only of ECS tasks with a single container.

(The stats for a single network interface _can_ vary slightly among
different containers within a single task stats API response, but I
believe that's only due to the ECS agent querying the stats from docker
for each container at slightly different times.)

So, there is a fully redundant label dimension right now. Remove it.

Signed-off-by: Ian Kerins <git@isk.haus>