Fix duplicate metrics for network stats and panics on invalid HTTP status codes (#98)
* Fix panic on invalid HTTP response status code

Turns out that the ECS task stats API can return HTTP 500 occasionally
on task initialization, and that we were failing to handle such
responses correctly by slicing out of bounds.

Remove the bad attempt at abbreviation. I think it was a mistake to add
in the first place.

Signed-off-by: Ian Kerins <git@isk.haus>

* Fix duplicate metrics for network stats

Turns out that it is in fact an error to report metrics that have the
same labels multiple times in a single `Collect` invocation - you do not
get "last wins" semantics. Deduplicate network metrics before reporting
them.

I failed to discover this in my synthetic testing because there was only
one container. I will fix that going forward.

Signed-off-by: Ian Kerins <git@isk.haus>

---------

Signed-off-by: Ian Kerins <git@isk.haus>