Add tooling to update task metadata API fixtures
The task metadata API fixtures I've previously created in this
repository were sourced from real ECS tasks managed by AWS CDK. But how
to reproduce that was left as an exercise to the reader.

Here, I am adding a CDK app and instructions that will enable anybody
with an AWS account and the willingness to install NodeJS and some other
tools on their computer to update the fixtures.

I have updated the fixtures by following the instructions I added to the
README. There are few substantive changes from the previous fixtures,
but there is a fair amount of churn, much of which is
unavoidable (timestamps, autoincrementing version numbers, etc), but
some of which can be eliminated through some new simple sorting of the
API responses, which should reduce churn going forward.

Signed-off-by: Ian Kerins <git@isk.haus>